Update AMD F14 Agesa to support Rev C0 cpus
This change updates the AMD Agesa code to support the Family 14
rev C0 cpus.  It also fixes (again) a ton of warnings, although
not all of them are gone.  The warning fixes affect code in the
Family 12 tree as well, so there are some small changes therein.
This code has been tested on a Persimmon and passes Abuild.
This is the first (and largest) of a number of commits to complete
the upgrade.

+++ src/vendorcode/amd/agesa/f14/Proc/GNB/Modules/GnbPcieInitLibV1/PcieSiliconServices.c
@@ -185,7 +185,7 @@ PcieSiliconUnHidePorts (
     Silicon->Address.AddressValue | D0F0x60_ADDRESS,
     D0F0x64_x0C_ADDRESS | IOC_WRITE_ENABLE,
     AccessS3SaveWidth32,
-    ~(BIT2 | BIT3 | BIT4 | BIT5 | BIT6 | BIT7),
+    ~(UINT32)(BIT2 | BIT3 | BIT4 | BIT5 | BIT6 | BIT7),
     0x0,
     GnbLibGetHeader (Pcie)
     );
@@ -193,7 +193,7 @@ PcieSiliconUnHidePorts (
     Silicon->Address.AddressValue | D0F0x60_ADDRESS,
     D0F0x64_x00_ADDRESS | IOC_WRITE_ENABLE,
     AccessS3SaveWidth32,
-    ~BIT6,
+    ~(UINT32)BIT6,
     BIT6,
     GnbLibGetHeader (Pcie)
     );
@@ -239,7 +239,7 @@ PcieSiliconHidePorts (
     Silicon->Address.AddressValue | D0F0x60_ADDRESS,
     D0F0x64_x0C_ADDRESS | IOC_WRITE_ENABLE,
     AccessS3SaveWidth32,
-    ~(BIT2 | BIT3 | BIT4 | BIT5 | BIT6 | BIT7),
+    ~(UINT32)(BIT2 | BIT3 | BIT4 | BIT5 | BIT6 | BIT7),
     D0F0x64_x0C.Value,
     GnbLibGetHeader (Pcie)
     );
@@ -247,7 +247,7 @@ PcieSiliconHidePorts (
     Silicon->Address.AddressValue | D0F0x60_ADDRESS,
     D0F0x64_x00_ADDRESS | IOC_WRITE_ENABLE,
     AccessS3SaveWidth32,
-    ~BIT6,
+    ~(UINT32)BIT6,
     0x0,
     GnbLibGetHeader (Pcie)
     );
